feat: Add viewport component + improve center-* components
This commit adds a new viewport component that sets its
width and height to 100vw and 100vh.
Also, this commit makes it so that flex is used in order
to center elements horizontally and vertically as opposed
to margin: auto. The benefit of this implementation is that
it allows both component to easily support multiple nested
elements.